+++ NEWS.md
@@ -1,5 +1,7 @@
 # spex dev
 
+* Now return all cells as polygons if raster has no values. 
+
 * Improved handling of `na.rm`, also now TRUE by default. #13 Thanks to Adriano Fantini.
 
 * A modest speed-up (up to 2X) by more careful use of R.